+++ examples/example-antd-management-fast-framework/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 1.9.21
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-playground@1.0.152
+  - antd-management-fast-design-react@1.0.94
+  - antd-management-fast-component@2.7.12
+  - antd-management-fast-framework@2.12.12
+  - antd-management-fast-formily@1.0.104
+  - antd-management-fast-common@2.6.125
+  - antd-management-fast-flow@1.0.94
+
 ## 1.9.20
 
 ### Patch Changes