+++ packages/antd-management-fast-flow/CHANGELOG.md
@@ -1,5 +1,14 @@
 # antd-management-fast-flow
 
+## 1.0.94
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-component@2.7.12
+  - antd-management-fast-common@2.6.125
+
 ## 1.0.93
 
 ### Patch Changes