+++ packages/antd-management-fast-formily/CHANGELOG.md
@@ -1,5 +1,15 @@
 # antd-management-fast-formily
 
+## 1.0.104
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - antd-management-fast-design-react@1.0.94
+  - antd-management-fast-component@2.7.12
+  - antd-management-fast-common@2.6.125
+
 ## 1.0.103
 
 ### Patch Changes